delete cluster attachment

+++ app/Http/Controllers/User/InfoController.php
@@ -9,18 +9,26 @@ use App\Models\TestHistory;
 use App\Models\TrainingPlan;
 use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Hash;
+use App\Http\Requests\User\UserChangePasswordRequest;
+use App\Http\Requests\User\InfoUpdateRequest;
 
 class InfoController extends Controller
 {
     public function info($id)
     {
+        if (Auth::user()->id != $id) {
+            return redirect()->back();
+        }
         $data["user"] = User::find($id);
         return view('user.info.info')->with('data', $data);
     }
 
-    public function update(Request $request)
+    public function update(InfoUpdateRequest $request,$id)
     {
-        $data["user"] = User::find($request->id);
+        if (Auth::user()->id != $id) {
+            return redirect()->back();
+        }
+        $data["user"] = User::find($id);
         $data["user"]->name = $request->name;
         $data["user"]->email = $request->email;
         $data["user"]->save();
@@ -31,41 +39,67 @@ class InfoController extends Controller
             $data["user"]->avatar = 'images/avatar/' . $data["user"]->id . '.' . $extension;
             $data["user"]->update();
         }
-        return redirect()->route('user.info.info')->with('data',$data)->with('profileChangeSuccess', 'Profile changed successfully');
+        return redirect()->route('user.info.info',$id)->with('data',$data)->with('success', 'Cập nhật thông tin thành công');
     }
 
     public function password($id)
     {
-        return view("user.info.password")->with("id", $id);
+        if (Auth::user()->id != $id) {
+            return redirect()->back();
+        }
+        $data["user"] = User::find($id);
+        return view("user.info.password")->with("id", $id)->with("data",$data);
     }
 
-    public function changePassword(Request $request)
+    public function changePassword(UserChangePasswordRequest $request,$id)
     {
-        $data["user"] = User::find($request->id);
-        $data["user"]->password = Hash::make($request->password);
-        $check = $data["user"]->save();
-        if ($check) {
-            return redirect()->back()->with('id', $request->id)->with("success", "Bạn đã thay đổi mật khẩu thành công");
+        if (Auth::user()->id != $id) {
+            return redirect()->back();
         }
-        return redirect()->back()->with('id', $request->id)->with("fail", "Đã có sự cố xảy ra.");
+        $data["user"] = User::find($id);
+        // $old_pasword = Hash::make($request->pastpass);
+        if(Hash::check($request->pastpass,$data["user"]->password)) {
+            $data["user"]->password = Hash::make($request->password);
+            $check = $data["user"]->save();
+        }
+        else {
+            return redirect()->back()->with("fail", "Mật khẩu cũ không đúng.");
+        }
+        
+        if ($check) {
+            return redirect()->back()->with('id', $id)->with("success", "Bạn đã thay đổi mật khẩu thành công");
+        }
+        return redirect()->back()->with('id', $id)->with("fail", "Đã có sự cố xảy ra.");
     }
 
     public function history($id)
     {
+        if (Auth::user()->id != $id) {
+            return redirect()->back();
+        }
+        $data["user"] = User::find($id);
         $hquery = TestHistory::where('user_id', '=', $id)
             ->orderBy('created_at', 'desc');
         $data["histories_count"] = $hquery->count();
-        $data["histories"] = $hquery->paginate(6);
+        $data["histories"] = $hquery->paginate(4);
         return view('user.info.history')->with('data',$data);
     }
 
     public function plan($id) {
+        if (Auth::user()->id != $id) {
+            return redirect()->back();
+        }
+        $data["user"] = User::find($id);
         $data["plan"] = TrainingPlan::where("user_id" ,'=',$id)->first();
         return view('user.info.plan')->with('data',$data);
     }
 
-    public function createPlan(Request $request) {
+    public function createPlan(Request $request,$id) {
+        if (Auth::user()->id != $id) {
+            return redirect()->back();
+        }
         $data["plan"] = new TrainingPlan();
+        $data["plan"]->user_id = $id;
         $data["plan"]->current_score = $request->current_score;
         $data["plan"]->goal_score = $request->goal_score;
         $data["plan"]->date_end = $request->date_end;
@@ -75,7 +109,24 @@ class InfoController extends Controller
         return redirect()->back()->with('data',$data)->with('success',"Bạn đã tạo kế hoạch thành công");
     }
 
+    public function updatePlan(Request $request,$id,$pid) {
+        if (Auth::user()->id != $id) {
+            return redirect()->back();
+        }
+        $data["plan"] = TrainingPlan::find($pid);
+        $data["plan"]->current_score = $request->current_score;
+        $data["plan"]->goal_score = $request->goal_score;
+        $data["plan"]->date_end = $request->date_end;
+        $data["plan"]->part_suggestion = $request->part_suggestion;
+        $data["plan"]->status = "ongoing";
+        $data["plan"]->save();
+        return redirect()->back()->with('data',$data)->with('success',"Bạn đã cập nhật kế hoạch thành công");
+    }
+
     public function deletePlan($id) {
+        if (Auth::user()->id != $id) {
+            return redirect()->back();
+        }
         $plan = TrainingPlan::find($id);
         $plan->delete();
         $id = Auth::user()->id;
